# frontend/lib/api/openapi.yml
openapi: 3.0.3
info:
  title: Kuyash Place Restaurant API
  version: 1.0.0
  description: |-
    Backend API for Kuyash Place Restaurant.

    **All money values are integer kobo** wrapped in a `{amount, currency, display}` object. Clients render `display` and never perform arithmetic on `amount`.
paths:
...
  /api/v1/orders/:
    get:
      operationId: orders_list
      description: |-
        Place an order, or list the caller's own.

        ``GET`` returns exactly what ``/orders/mine/`` returns. The route and this
        class have always been named for both verbs, and the scoped throttle
        deliberately counts writes only so that reading your own orders cannot use
        up the allowance for placing one — but only ``post`` was ever implemented,
        so listing answered 405.
      summary: List my orders
      parameters:
      - name: cursor
        required: false
        in: query
        description: The pagination cursor value.
        schema:
          type: string
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - orders
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrderListList'
          description: ''
    post:
      operationId: orders_create
      description: |-
        Place an order, or list the caller's own.

        ``GET`` returns exactly what ``/orders/mine/`` returns. The route and this
        class have always been named for both verbs, and the scoped throttle
        deliberately counts writes only so that reading your own orders cannot use
        up the allowance for placing one — but only ``post`` was ever implemented,
        so listing answered 405.
      summary: Place an order
      parameters:
      - in: header
        name: Idempotency-Key
        schema:
          type: string
        description: A UUID. Replaying it returns the original response.
        required: true
      tags:
      - orders
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceOrderRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PlaceOrderRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PlaceOrderRequest'
        required: true
      security:
      - cookieAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
          description: ''
...
  /api/v1/orders/{reference}/:
    get:
      operationId: orders_retrieve
      description: One order. The polling target.
      summary: Retrieve an order
      parameters:
      - in: header
        name: If-None-Match
        schema:
          type: string
        description: Send the previous ETag to receive 304 when unchanged.
      - in: header
        name: X-Guest-Token
        schema:
          type: string
        description: Required for guest orders.
      - in: path
        name: reference
        schema:
          type: string
        required: true
      tags:
      - orders
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
          description: ''
...
  /api/v1/kds/orders/:
    get:
      operationId: kds_orders_retrieve
      description: The live ticket queue, oldest first.
      summary: Kitchen queue
      parameters:
      - in: query
        name: status
        schema:
          type: string
        description: Comma-separated statuses.
      tags:
      - kds
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KDSQueue'
          description: ''
...
  /api/v1/kds/orders/{reference}/{action}/:
    post:
      operationId: kds_orders_create
      description: Accept, reject or advance a ticket.
      summary: Advance a ticket
      parameters:
      - in: path
        name: action
        schema:
          type: string
        required: true
      - in: path
        name: reference
        schema:
          type: string
        required: true
      tags:
      - kds
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KDSTicket'
          description: ''
...
components:
  schemas:
...
    KDSQueue:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/KDSTicket'
        reject_reasons:
          type: array
          items:
            $ref: '#/components/schemas/KDSRejectReason'
      required:
      - orders
      - reject_reasons
    KDSRejectReason:
      type: object
      properties:
        code:
          type: string
        title:
          type: string
      required:
      - code
      - title
...
    KDSTicket:
      type: object
      description: One kitchen ticket.
      properties:
        reference:
          type: string
        status:
          type: string
        placed_at:
          type: string
          format: date-time
        elapsed_seconds:
          type: integer
        is_late:
          type: boolean
        fulfilment_type:
          type: string
        zone:
          type: string
        payment_status:
          type: string
        payment_method:
          type: string
        requires_cash_collection:
          type: boolean
        rider:
          allOf:
          - $ref: '#/components/schemas/KDSTicketRider'
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/KDSTicketLine'
        customer_note:
          type: string
        customer:
          $ref: '#/components/schemas/KDSTicketCustomer'
        grand_total:
          $ref: '#/components/schemas/Money'
      required:
      - customer
      - customer_note
      - elapsed_seconds
      - fulfilment_type
      - grand_total
      - is_late
      - items
      - payment_method
      - payment_status
      - placed_at
      - reference
      - requires_cash_collection
      - rider
      - status
      - zone
    KDSTicketCustomer:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
      required:
      - name
      - phone
    KDSTicketLine:
      type: object
      properties:
        quantity:
          type: integer
        name:
          type: string
        variant:
          type: string
        modifiers:
          type: array
          items:
            type: string
        special_instructions:
          type: string
      required:
      - modifiers
      - name
      - quantity
      - special_instructions
      - variant
    KDSTicketRider:
      type: object
      properties:
        name:
          type: string
      required:
      - name
...
    Money:
      type: object
      description: |-
        Read-only representation of a money amount.

        Always emits ``{amount, currency, display}``. ``amount`` is integer kobo;
        ``display`` is the formatted string clients render. Clients never format
        currency themselves and never do arithmetic on ``amount``.
      properties:
        amount:
          type: integer
          readOnly: true
          description: Integer kobo. ₦1.00 = 100.
        currency:
          type: string
          readOnly: true
          default: NGN
        display:
          type: string
          readOnly: true
          description: Formatted, e.g. ₦12,500.00
      required:
      - amount
      - currency
      - display
...
    OrderAddress:
      type: object
      properties:
        recipient_name:
          type: string
        phone:
          type: string
        street:
          type: string
        area:
          type: string
        city:
          type: string
        state:
          type: string
        landmark:
          type: string
        delivery_notes:
          type: string
        zone:
          type: string
      required:
      - area
      - city
      - delivery_notes
      - landmark
      - phone
      - recipient_name
      - state
      - street
      - zone
    OrderDetail:
      type: object
      description: Response shape for order detail and the polling endpoint.
      properties:
        reference:
          type: string
        status:
          type: string
        status_display:
          type: string
        payment_status:
          type: string
        payment_method:
          type: string
        fulfilment_type:
          type: string
        placed_at:
          type: string
          format: date-time
          nullable: true
        estimated_ready_at:
          type: string
          format: date-time
          nullable: true
        estimated_delivery_at:
          type: string
          format: date-time
          nullable: true
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/OrderTimelineStep'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderLine'
        delivery_address:
          allOf:
          - $ref: '#/components/schemas/OrderAddress'
          nullable: true
        totals:
          $ref: '#/components/schemas/Totals'
        promo_code:
          type: string
        customer_note:
          type: string
        can_cancel:
          type: boolean
        rider:
          allOf:
          - $ref: '#/components/schemas/OrderRider'
          nullable: true
        guest_token:
          type: string
      required:
      - can_cancel
      - customer_note
      - delivery_address
      - estimated_delivery_at
      - estimated_ready_at
      - fulfilment_type
      - items
      - payment_method
      - payment_status
      - placed_at
      - promo_code
      - reference
      - rider
      - status
      - status_display
      - timeline
      - totals
    OrderLine:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        variant_name:
          type: string
        quantity:
          type: integer
        modifiers:
          type: array
          items:
            type: string
        special_instructions:
          type: string
        unit_price:
          $ref: '#/components/schemas/Money'
        line_subtotal:
          $ref: '#/components/schemas/Money'
        image_url:
          type: string
          nullable: true
        review:
          allOf:
          - $ref: '#/components/schemas/OrderLineReview'
          nullable: true
          description: The customer's review of this line, if they wrote one.
        can_review:
          type: boolean
      required:
      - can_review
      - id
      - image_url
      - line_subtotal
      - modifiers
      - name
      - quantity
      - review
      - slug
      - special_instructions
      - unit_price
      - variant_name
...
    OrderList:
      type: object
      description: Compact shape for order history.
      properties:
        reference:
          type: string
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/OrderStatusEnum'
          readOnly: true
        status_display:
          type: string
          readOnly: true
        payment_status:
          allOf:
          - $ref: '#/components/schemas/PaymentStatusEnum'
          readOnly: true
        fulfilment_type:
          type: string
          readOnly: true
        placed_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        total:
          allOf:
          - $ref: '#/components/schemas/Money'
          readOnly: true
        item_count:
          type: integer
          readOnly: true
        preview:
          type: array
          items:
            $ref: '#/components/schemas/OrderPreviewLine'
          readOnly: true
      required:
      - fulfilment_type
      - item_count
      - payment_status
      - placed_at
      - preview
      - reference
      - status
      - status_display
      - total
...
    OrderRider:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
      required:
      - name
      - phone
    OrderStatusEnum:
      enum:
      - pending_payment
      - paid
      - confirmed
      - preparing
      - ready
      - out_for_delivery
      - delivered
      - rejected
      - cancelled
      - expired
      - failed_delivery
      - refunded
      - failed
      type: string
      description: |-
        * `pending_payment` - Awaiting payment
        * `paid` - Paid
        * `confirmed` - Confirmed by the kitchen
        * `preparing` - Preparing
        * `ready` - Ready
        * `out_for_delivery` - Out for delivery
        * `delivered` - Delivered
        * `rejected` - Rejected
        * `cancelled` - Cancelled
        * `expired` - Expired
        * `failed_delivery` - Delivery failed
        * `refunded` - Refunded
        * `failed` - Failed
    OrderTimelineStep:
      type: object
      properties:
        status:
          type: string
        at:
          type: string
          format: date-time
          nullable: true
        reached:
          type: boolean
        label:
          type: string
      required:
      - at
      - label
      - reached
      - status
...
    PaginatedOrderListList:
      type: object
      required:
      - results
      properties:
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cD00ODY%3D"
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cj0xJnA9NDg3
        results:
          type: array
          items:
            $ref: '#/components/schemas/OrderList'
...
    PlaceOrderPaymentMethodEnum:
      enum:
      - card
      - transfer
      - cash
      type: string
      description: |-
        * `card` - card
        * `transfer` - transfer
        * `cash` - cash
    PlaceOrderRequest:
      type: object
      description: |-
        Input for placing an order.

        Contains no price field of any kind. ``expected_total`` is a *guard*, not an
        instruction: if it disagrees with the server's figure the order is refused
        rather than silently charged at a different amount.
      properties:
        payment_method:
          $ref: '#/components/schemas/PlaceOrderPaymentMethodEnum'
        payment_provider:
          type: string
          default: ''
        customer_note:
          type: string
          default: ''
          maxLength: 500
        expected_total:
          type: integer
          nullable: true
          description: The grand total the customer was shown, in kobo.
        guest:
          type: object
          additionalProperties: {}
      required:
      - payment_method
...
    Totals:
      type: object
      description: |-
        The money breakdown shared by carts and orders.

        Declared so the generated schema types every line as money instead of an
        untyped dictionary — the frontend renders each ``display`` and has nothing
        to add up.
      properties:
        subtotal:
          $ref: '#/components/schemas/Money'
        discount:
          $ref: '#/components/schemas/Money'
        delivery_fee:
          $ref: '#/components/schemas/Money'
        service_charge:
          $ref: '#/components/schemas/Money'
        vat:
          $ref: '#/components/schemas/Money'
        tip:
          $ref: '#/components/schemas/Money'
        grand_total:
          $ref: '#/components/schemas/Money'
      required:
      - delivery_fee
      - discount
      - grand_total
      - service_charge
      - subtotal
      - tip
      - vat
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: kuyash_session